+++ crates/napi/src/bindgen_runtime/js_values/promise_raw.rs
@@ -64,6 +64,42 @@ impl<T> PromiseRaw<'_, T> {
   }
 }
 
+impl<'env, T: ToNapiValue> PromiseRaw<'env, T> {
+  /// Create a new promise and resolve it with the given value
+  pub fn resolve(env: &Env, value: T) -> Result<Self> {
+    let mut deferred = ptr::null_mut();
+    let mut promise = ptr::null_mut();
+    check_status!(
+      unsafe { sys::napi_create_promise(env.0, &mut deferred, &mut promise) },
+      "Failed to create promise"
+    )?;
+    check_status!(
+      unsafe {
+        sys::napi_resolve_deferred(env.0, deferred, ToNapiValue::to_napi_value(env.0, value)?)
+      },
+      "Failed to resolve promise"
+    )?;
+    Ok(PromiseRaw::new(env.0, promise))
+  }
+
+  /// Create a new promise and reject it with the given error
+  pub fn reject<E: ToNapiValue>(env: &Env, error: E) -> Result<Self> {
+    let mut deferred = ptr::null_mut();
+    let mut promise = ptr::null_mut();
+    check_status!(
+      unsafe { sys::napi_create_promise(env.0, &mut deferred, &mut promise) },
+      "Failed to create promise"
+    )?;
+    check_status!(
+      unsafe {
+        sys::napi_reject_deferred(env.0, deferred, ToNapiValue::to_napi_value(env.0, error)?)
+      },
+      "Failed to reject promise"
+    )?;
+    Ok(PromiseRaw::new(env.0, promise))
+  }
+}
+
 impl<'env, T: FromNapiValue> PromiseRaw<'env, T> {
   /// Promise.then method
   pub fn then<'then, Callback, U>(&self, cb: Callback) -> Result<PromiseRaw<'env, U>>