+++ crates/macro/CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.4.0](https://github.com/napi-rs/napi-rs/compare/napi-derive-v3.3.3...napi-derive-v3.4.0) - 2025-12-02
+
+### Added
+
+- *(napi-derive)* add tracing feature for debug logging NAPI function calls ([#3041](https://github.com/napi-rs/napi-rs/pull/3041))
+
+### Fixed
+
+- *(deps)* update rust crate convert_case to 0.10 ([#3031](https://github.com/napi-rs/napi-rs/pull/3031))
+
 ## [3.3.3](https://github.com/napi-rs/napi-rs/compare/napi-derive-v3.3.2...napi-derive-v3.3.3) - 2025-11-10
 
 ### Fixed